requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
django==6.0; python_version >= '3.12'
djangorestframework==3.16.1; python_version >= '3.9'
sqlparse==0.5.5; python_version >= '3.8'
